feat, refactor: make use of new parser for nicer cmp preview docs
Now the parser introduced in #20 is utilized to provide more info to
cmp items. This info is used to improve the cmp documentation / preview
window (currently defaults to a hardcoded implementation showing
the description and snippet contents).

Snippets with option 'r' or 'e' are parsed but not passed on to cmp
at the moment.